fixed problem with laptop setup

+++ config_management/install_configs.sh
@@ -7,13 +7,17 @@ sudo cp $DIR/etc/pacman.conf /etc
 
 sudo cp $DIR/etc/X11/xorg.conf.d/00-keyboard.conf /etc/X11/xorg.conf.d
 sudo cp $DIR/etc/X11/xorg.conf.d/01-mouse.conf /etc/X11/xorg.conf.d
-[[ -e /etc/X11/xorg.conf.d/10-trackpoint.conf ]] && cp /etc/X11/xorg.conf.d/10-trackpoint.conf $DIR/etc/X11/xorg.conf.d
-[[ -e /etc/X11/xorg.conf.d/11-touchpad.conf ]] && cp /etc/X11/xorg.conf.d/11-touchpad.conf $DIR/etc/X11/xorg.conf.d
-[[ -e $HOME/.xbindkeysrc ]] && cp $DIR/home/xbindkeysrc $HOME/.xbindkeysrc
 
-mkdir $HOME/.config
+read -n1 -p "Include laptop configs? [y/n] " laptop
+if [[ "$laptop" == "y" ]]; then
+   sudo cp $DIR/etc/X11/xorg.conf.d/10-trackpoint.conf /etc/X11/xorg.conf.d
+   sudo cp $DIR/etc/X11/xorg.conf.d/11-touchpad.conf /etc/X11/xorg.conf.d
+   [[ -e $HOME/.xbindkeysrc ]] && cp $DIR/home/xbindkeysrc $HOME/.xbindkeysrc
+fi
+
+[[ ! -d "$HOME/.config" ]] && mkdir $HOME/.config
 cp $DIR/home/config/user-dirs.dirs $HOME/.config
-git clone https://github.com/chriskempson/base16-shell.git ~/.config/base16-shell
+[[ ! -d "$HOME/.config/base16-shell" ]] && git clone https://github.com/chriskempson/base16-shell.git ~/.config/base16-shell
 
 cp $DIR/home/bash_profile $HOME/.bash_profile
 cp $DIR/home/xinitrc $HOME/.xinitrc
